HOSPITAL_UAS/requirements.txt
<<<<<<< Updated upstream
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.3
greenlet==3.2.4
h11==0.16.0
idna==3.11
numpy==2.3.5
pandas==2.3.3
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
# Dependensi yang Diperbarui: passlib dengan bcrypt sebagai dependensi opsional
passlib
bcrypt==4.0.1
fastapi
uvicorn
sqlalchemy
pydantic
pandas
pymysql
pytest
httpx
=======
fastapi
uvicorn
pandas
python-dotenv
>>>>>>> Stashed changes
